rag_local/phase1-indexing/requirements.txt
sentence-transformers==3.0.0
langchain-text-splitters==0.2.0
pypdf==4.2.0
python-docx==1.1.2
psycopg2-binary==2.9.9
elasticsearch==8.13.0
weaviate-client==4.6.4
tiktoken==0.7.0